// prisma/generated/models/ApiCallBilling.ts
export type ApiCallBillingUncheckedCreateInput = {
  id?: number
  eventId: string
  userId: number
  model: string
  inputToken?: number
  outputToken?: number
  cost?: runtime.Decimal | runtime.DecimalJsLike | number | string
  walletId: number
  status?: number
  durationMs?: number
  timestamp: Date | string
  processedAt?: Date | string | null
  isBilled?: boolean
  errorMessage?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  clientIp?: string
  externalTraceId?: string
}